Start uniquify.
5to6-codemod isn't de-clashing the case of module.exports having
a property set to a top level declaration. This causes a clash. For
example:
`module.exports = { myDeclaration }`
Would still let the current "myDeclaration" declaration exist, while
also trying to write an ExportNamedDeclaration that will clash,
re-defining the existing declaration. This new plugin proposes to
re-write the file to detect this clash, & rename the existing
declaration and all references to it.

No implementation at this stage, just skeleton.